Move logic into SAR service from controller.
