+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppssubjectaccessrequestapi/gateways/SubjectAccessRequestGateway.kt
@@ -27,4 +27,7 @@ class SubjectAccessRequestGateway(
 
     return response
   }
+  fun saveSubjectAccessRequest() {
+
+  }
 }
